WIP major refactoring.
