+++ configurator/Cargo.toml
@@ -9,6 +9,8 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0.40"
 bytes = "1.0.1"
+mockall = "0.9.1"
+mockall_double = "0.2.0"
 prost = "0.7.0"
 serialport = "4.0.1"
